refactor: Extract fixtures file
