+++ packages/cozy-harvest-lib/test/fixtures.js
@@ -1,0 +1,159 @@
+
+const fixtures = {
+  credentials: {
+    username: 'foo',
+    passphrase: 'bar'
+  },
+  konnector: {
+    slug: 'konnectest',
+    fields: {
+      username: {
+        type: 'text'
+      },
+      passphrase: {
+        type: 'password'
+      }
+    }
+  },
+  konnectorWithFolder: {
+    _type: 'io.cozy.konnectors',
+    name: 'myBills',
+    slug: 'mybills',
+    fields: {
+      advancedFields: {
+        folderPath: {
+          advanced: true
+        }
+      }
+    }
+  },
+  folder: {
+    _id: '3f5b288af36041f189ec22063adab706'
+  },
+  folderPath: '/Administrative/myBills/foo',
+  folderPermission: {
+    saveFolder: {
+      type: 'io.cozy.files',
+      values: ['3f5b288af36041f189ec22063adab706'],
+      verbs: ['GET', 'PATCH', 'POST']
+    }
+  },
+  triggerAttributes: {
+    arguments: '0 0 0 * * 0',
+    type: '@cron',
+    worker: 'konnector',
+    message: {
+      account: 'created-account-id',
+      konnector: 'konnectest'
+    }
+  },
+  account: {
+    account_type: 'konnectest',
+    auth: {
+      username: 'foo',
+      passphrase: 'bar'
+    },
+    identifier: 'username'
+  },
+  createdAccount: {
+    _id: 'created-account-id',
+    account_type: 'konnectest',
+    auth: {
+      username: 'foo',
+      passphrase: 'bar'
+    },
+    identifier: 'username'
+  },
+  updatedAccount: {
+    _id: 'updated-account-id',
+    account_type: 'konnectest',
+    auth: {
+      username: 'foo',
+      passphrase: 'fuz'
+    }
+  },
+  existingAccount: {
+    _id: 'existing-account-id',
+    account_type: 'konnectest',
+    auth: {
+      username: 'foo',
+      passphrase: 'bar'
+    },
+    identifier: 'username'
+  },
+  existingTrigger: {
+    id: 'existing-trigger-id',
+    _type: 'io.cozy.triggers',
+    attributes: {
+      arguments: '0 0 0 * * 0',
+      type: '@cron',
+      worker: 'konnector',
+      message: {
+        account: 'existing-account-id',
+        konnector: 'konnectest'
+      }
+    }
+  },
+  bankingKonnectorAccountAttributes: {
+    auth: {
+      identifier: '80564789',
+      secret: '13337'
+    },
+    identifier: '80564789'
+  },
+  createdTrigger: {
+    id: 'created-trigger-id',
+    _type: 'io.cozy.triggers',
+    attributes: {
+      arguments: '0 0 0 * * 0',
+      type: '@cron',
+      worker: 'konnector',
+      message: {
+        account: 'updated-account-id',
+        konnector: 'konnectest'
+      }
+    }
+  },
+  launchedJob: {
+    type: 'io.cozy.jobs',
+    _id: 'lauched-job-id',
+    id: 'lauched-job-id',
+    domain: 'cozy.tools:8080',
+    worker: 'konnector',
+    state: 'running',
+    queued_at: '2016-09-19T12:35:08Z',
+    started_at: '2016-09-19T12:35:08Z',
+    error: '',
+    links: {
+      self: '/jobs/lauched-job-id'
+    }
+  },
+  runningJob: {
+    type: 'io.cozy.jobs',
+    id: 'running-job-id',
+    domain: 'cozy.tools:8080',
+    worker: 'konnector',
+    state: 'running',
+    queued_at: '2016-09-19T12:35:08Z',
+    started_at: '2016-09-19T12:35:08Z',
+    error: '',
+    links: {
+      self: '/jobs/running-job-id'
+    }
+  },
+  doneJob: {
+    type: 'io.cozy.jobs',
+    id: 'done-job-id',
+    domain: 'cozy.tools:8080',
+    worker: 'konnector',
+    state: 'done',
+    queued_at: '2016-09-19T12:35:08Z',
+    started_at: '2016-09-19T12:35:08Z',
+    error: '',
+    links: {
+      self: '/jobs/done-job-id'
+    }
+  }
+}
+
+export default fixtures
